chart: set samesite for forward-auth
This sets the `SameSite` cookie attribute to `None` for the
forward-auth deployment.

We do this to prevent redirect loops when upstream services
handle their own authentication flow (e.g. using SAML) where the
callback request from the identity provider is an HTTP POST.

Such requests are by definition cross-site. The default `SameSite`
value of `Lax` results in the session cookie not being included.
This essentially means that the user is considered unauthenticated by
the forward-auth service when the callback request is processed.