ci: allow downloading newer toolchain
The default images for Go sets GOTOOLCHAIN=local
which forces builds to use the bundled Go version.

Setting this to auto allows for upgrading to a newer version
if specified by the toolchain directive in go.mod.

Useful for situations where the latest Go Docker image lags
behind the latest Go version.